Rawa_back/requirements.txt

Django==3.2
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==9.4.0
psycopg2-binary==2.9.5
django-filter==2.4.0
channels
channels-redis
